feat: Reformat code and separate variable parsing
